+++ packages/connect/src/base/listener.ts
@@ -46,9 +46,7 @@ enum ListenerState {
 
 type Address = { port: number; address: string }
 
-type NATSituation =
-  | { bidirectionalNAT: true }
-  | { bidirectionalNAT: false; externalAddress: string; externalPort: number; isExposed: boolean }
+type NATSituation = { bidirectionalNAT: boolean; externalAddress?: string; externalPort?: number; isExposed?: boolean }
 
 export type ProtocolListener = {
   identifier: string
@@ -247,14 +245,19 @@ class Listener extends EventEmitter<ListenerEvents> implements InterfaceListener
 
     this.attachSocketHandlers()
 
-    const natSituation = await this.checkNATSituation(ownInterface.address, ownInterface.port)
+    const natSituation: NATSituation = await this.checkNATSituation(ownInterface.address, ownInterface.port)
 
     log(`NAT situation detected: `, natSituation)
-    const internalInterfaces = getAddrs(ownInterface.port, {
+
+    // Only expose localhost and private address to the DHT if the node is
+    // configured to also accept connections for these.
+    const addressOptions = {
       useIPv4: true,
-      includePrivateIPv4: true,
-      includeLocalhostIPv4: true
-    })
+      includePrivateIPv4: this.options.allowPrivateConnections,
+      includeLocalhostIPv4: this.options.allowLocalConnections
+    }
+
+    const internalInterfaces = getAddrs(ownInterface.port, addressOptions)
 
     if (!natSituation.bidirectionalNAT && natSituation.isExposed) {
       // If any of the interface addresses is the
@@ -267,8 +270,8 @@ class Listener extends EventEmitter<ListenerEvents> implements InterfaceListener
 
       this.addrs.external = [
         nodeToMultiaddr({
-          address: natSituation.externalAddress,
-          port: natSituation.externalPort,
+          address: natSituation.externalAddress as string,
+          port: natSituation.externalPort as number,
           family: 'IPv4'
         })
       ]
@@ -282,19 +285,20 @@ class Listener extends EventEmitter<ListenerEvents> implements InterfaceListener
     this.connectComponents.getRelay().start()
 
     this._emitListening()
+    this.state = ListenerState.LISTENING
 
-    // If node is supposed to announce with routable address -> don't assign to other relays
-    // If node is running behind bidirectional NAT or deteced as not being exposed -> assign to other relays
-    // If node is not supposed to announce with routable address -> assign to other relays as fallback
-    if (!this.options.announce || natSituation.bidirectionalNAT || !natSituation.isExposed) {
-      this.connectComponents.getEntryNodes().on(RELAY_CHANGED_EVENT, this._emitListening)
-
-      // Instructs entry node manager to assign to available
-      // entry once startup has finished
-      this.connectComponents.getEntryNodes().enable()
+    // If node is a PRN (the `--announce` flag is set), disable relay functionality
+    // If node is not supposed to have any relayed connection (the `--noRelay` flag is set), disable relay functionality
+    if (this.options.announce || this.options.noRelay) {
+      // skip PRN
+      return
     }
 
-    this.state = ListenerState.LISTENING
+    this.connectComponents.getEntryNodes().on(RELAY_CHANGED_EVENT, this._emitListening)
+
+    // Instructs entry node manager to assign to available
+    // entry once startup has finished
+    this.connectComponents.getEntryNodes().enable()
   }
 
   /**